Actualización del sistema escolar

+++ restaurar.py
@@ -4,29 +4,44 @@ from conexion import obtener_conexion
 
 def restaurar_tabla(nombre_tabla):
 
-    with open(f"respaldos/{nombre_tabla}.json", "r", encoding="utf-8") as archivo:
-        datos = json.load(archivo)
+    archivo_ruta = f"respaldos/{nombre_tabla}.json"
+
+    try:
+        with open(archivo_ruta, "r", encoding="utf-8") as archivo:
+            datos = json.load(archivo)
+
+    except FileNotFoundError:
+        print(f"No existe el respaldo de la tabla: {nombre_tabla}")
+        return
+
+    if len(datos) == 0:
+        print("El archivo de respaldo está vacío")
+        return
 
     conexion = obtener_conexion()
     cursor = conexion.cursor()
 
+    
+    cursor.execute("SET FOREIGN_KEY_CHECKS=0")
+
     cursor.execute(f"TRUNCATE TABLE {nombre_tabla}")
 
+    
     for fila in datos:
 
         columnas = ", ".join(fila.keys())
-        placeholders = ", ".join(["%s"] * len(fila))
+        valores = ", ".join(["%s"] * len(fila))
 
         sql = f"""
-        INSERT INTO {nombre_tabla}
-        ({columnas})
-        VALUES ({placeholders})
+        INSERT INTO {nombre_tabla} ({columnas})
+        VALUES ({valores})
         """
 
         cursor.execute(sql, tuple(fila.values()))
 
+    cursor.execute("SET FOREIGN_KEY_CHECKS=1")
+
     conexion.commit()
-
-    print(f"Tabla {nombre_tabla} restaurada")
-
     conexion.close()
+
+    print(f"Tabla '{nombre_tabla}' restaurada correctamente desde JSON")
